# packages/contracts/openapi/laundry-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "auth.pin_challenge.request": {
        "oneOf": [
          {
            "additionalProperties": false,
            "properties": {
              "purpose": {
                "const": "quick_switch",
                "type": "string"
              },
              "target_staff_id": {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              }
            },
            "required": [
              "purpose",
              "target_staff_id"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "approver_staff_id": {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              },
              "pending_action_ref": {
                "pattern": "^[\\x21-\\x7E]{1,256}$",
                "type": "string"
              },
              "purpose": {
                "const": "step_up",
                "type": "string"
              }
            },
            "required": [
              "purpose",
              "pending_action_ref",
              "approver_staff_id"
            ],
            "type": "object"
          }
        ]
      },
      "auth.pin_challenge.response": {
        "additionalProperties": false,
        "properties": {
          "challenge_id": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "expires_at": {
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "max_attempts": {
            "const": 5,
            "type": "number"
          },
          "purpose": {
            "enum": [
              "quick_switch",
              "step_up"
            ],
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "purpose",
          "expires_at",
          "max_attempts"
        ],
        "type": "object"
      },
      "auth.pin_verify.request": {
        "additionalProperties": false,
        "properties": {
          "challenge_id": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "pin": {
            "pattern": "^[0-9]{4,8}$",
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "pin"
        ],
        "type": "object"
      },
      "auth.pin_verify.response": {
        "anyOf": [
          {
            "additionalProperties": false,
            "properties": {
              "access_token": {
                "pattern": "^[A-Za-z0-9_-]+(?:\\.[A-Za-z0-9_-]+){2}$",
                "type": "string"
              },
              "display": {
                "additionalProperties": false,
                "properties": {
                  "org_code": {
                    "type": "string"
                  },
                  "staff_name": {
                    "type": "string"
                  },
                  "store_code": {
                    "type": "string"
                  },
                  "store_name": {
                    "type": "string"
                  }
                },
                "required": [
                  "store_name",
                  "staff_name",
                  "org_code",
                  "store_code"
                ],
                "type": "object"
              },
              "expires_in": {
                "const": 900,
                "type": "number"
              },
              "features": {
                "additionalProperties": {
                  "type": "boolean"
                },
                "type": "object"
              },
              "role": {
                "enum": [
                  "admin",
                  "staff"
                ],
                "type": "string"
              },
              "session": {
                "additionalProperties": false,
                "properties": {
                  "device_id": {
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                    "type": "string"
                  },
                  "org_id": {
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                    "type": "string"
                  },
                  "permission_version": {
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991,
                    "type": "integer"
                  },
                  "session_id": {
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                    "type": "string"
                  },
                  "session_version": {
                    "exclusiveMinimum": 0,
                    "maximum": 9007199254740991,
                    "type": "integer"
                  },
                  "staff_id": {
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                    "type": "string"
                  },
                  "store_id": {
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                    "type": "string"
                  }
                },
                "required": [
                  "session_id",
                  "session_version",
                  "org_id",
                  "store_id",
                  "staff_id",
                  "device_id",
                  "permission_version"
                ],
                "type": "object"
              },
              "storage": {
                "const": "memory_only",
                "type": "string"
              },
              "token_type": {
                "const": "Bearer",
                "type": "string"
              }
            },
            "required": [
              "access_token",
              "token_type",
              "expires_in",
              "storage",
              "session",
              "role",
              "features",
              "display"
            ],
            "type": "object"
          },
          {
            "additionalProperties": false,
            "properties": {
              "expires_at": {
                "maximum": 9007199254740991,
                "minimum": 0,
                "type": "integer"
              },
              "step_up_proof_id": {
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                "type": "string"
              }
            },
            "required": [
              "step_up_proof_id",
              "expires_at"
            ],
            "type": "object"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Memory-only access token (never store in cookies or Web Storage).",
        "scheme": "bearer",
        "type": "http"
      },
      "csrfHeader": {
        "description": "Double-submit CSRF header paired with the readable CSRF cookie.",
        "in": "header",
        "name": "x-csrf-token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Contract-first OpenAPI 3.1 projection for frozen M1 and M2 counter contracts. Auth paths come solely from AUTH_OPERATION_MATRIX. Bus commands/queries come from M1_FIRST_WAVE_DEFINITIONS and M2_CONTRACT_DEFINITIONS. Errors use the A2 CommandResponse / CommandFailureResponse envelope.",
    "title": "Laundry Desk v2 API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/auth/pin/challenges": {
      "post": {
        "description": "Browser auth operation `pin_challenge` (ingress: browser_session). Allowed public errors: VALIDATION_FAILED, AUTHENTICATION_FAILED, PERMISSION_DENIED, RESOURCE_UNAVAILABLE, CSRF_REJECTED, RATE_LIMITED, INVARIANT_FAILED, TRANSACTION_FAILED, EVENT_DISPATCH_FAILED. Projection source: AUTH_OPERATION_MATRIX (A5/A7). Server-only hashes and factories are omitted.",
        "operationId": "auth_pin_challenge",
        "parameters": [
          {
            "description": "Double-submit CSRF proof; must match the readable CSRF cookie value.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "pattern": "^v1\\.[A-Za-z0-9_-]{43,128}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth.pin_challenge.request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth.pin_challenge.response"
                }
              }
            },
            "description": "Successful browser auth response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Request origin verification failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Too many requests"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Unified A2 command failure envelope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "csrfHeader": []
          }
        ],
        "summary": "Auth pin_challenge",
        "tags": [
          "auth"
        ],
        "x-laundry-kind": "auth"
      }
    },
    "/api/v2/auth/pin/challenges/{challenge_id}/verify": {
      "post": {
        "description": "Browser auth operation `pin_verify` (ingress: browser_session). Allowed public errors: VALIDATION_FAILED, AUTHENTICATION_FAILED, PERMISSION_DENIED, RESOURCE_UNAVAILABLE, CSRF_REJECTED, RATE_LIMITED, INVARIANT_FAILED, TRANSACTION_FAILED, EVENT_DISPATCH_FAILED. Projection source: AUTH_OPERATION_MATRIX (A5/A7). Server-only hashes and factories are omitted.",
        "operationId": "auth_pin_verify",
        "parameters": [
          {
            "in": "path",
            "name": "challenge_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Double-submit CSRF proof; must match the readable CSRF cookie value.",
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "pattern": "^v1\\.[A-Za-z0-9_-]{43,128}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/auth.pin_verify.request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth.pin_verify.response"
                }
              }
            },
            "description": "Successful browser auth response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Request origin verification failed"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Too many requests"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandFailureResponse"
                }
              }
            },
            "description": "Unified A2 command failure envelope"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "csrfHeader": []
          }
        ],
        "summary": "Auth pin_verify",
        "tags": [
          "auth"
        ],
        "x-laundry-kind": "auth"
      }
    },
...
  }
}